// src/app/models/Order.ts
export class Order {
    orderId: number;
    buyerEmail: string;
    buyerName: string;
    buyerSurname: string;
    buyerPhone: string;
    buyerUlica: string;
    buyerBroj: string;
    buyerInterfon: string;
    buyerBrojStana: string;
    buyerSprat: string;
    orderAmount: string;
    orderStatus: string;
    createTime: string;
    updateTime: string;
    vreme: string;

}